- Fixture에서 더미데이터를 사용할 때 Mocking하고 싶었지만, 실제 Repository레이어를 사용하기 때문에 내부 서비스만 mocking

+++ src/test/java/camp/woowak/lab/fixture/CartFixture.java
@@ -25,7 +25,7 @@ public interface CartFixture {
 		return new TestStore(id, vendor, name, "송파", "010-1234-5678", minOrderPrice, startTIme, endTime);
 	}
 
-	default Menu createSavedMenu(Long id, Store store, MenuCategory menuCategory, String name, int price) {
+	default Menu createSavedMenu(Long id, Store store, MenuCategory menuCategory, String name, Long price) {
 		return new TestMenu(id, store, menuCategory, name, price);
 	}
 }
